Posting APIs (#4)
* generate json-friendly plan controller

* enable basic api routes for plans

* add option to view all plans outside of user scope

* enable basic api routes for garden beds

* add csv package

* use nimblecsv instead

* use float instead of integer for bed width and length

* parse uploaded file as CSV with bedcontroller
- add csv parser util

* check for bed and plant id before creating plan

* add query to get bed by soil type

* enable uploads in plan controller

* ensure plant names are unique in the database

+++ priv/repo/migrations/20250528223050_create_beds.exs
@@ -5,8 +5,8 @@ defmodule Planter.Repo.Migrations.CreateBeds do
     create table(:beds) do
       add :x, :integer
       add :y, :integer
-      add :width, :integer
-      add :length, :integer
+      add :width, :float
+      add :length, :float
       add :soil_type, :string
 
       timestamps(type: :utc_datetime)
