+++ priv/repo/migrations/20250528224156_create_plans.exs
@@ -3,7 +3,7 @@ defmodule Planter.Repo.Migrations.CreatePlans do
 
   def change do
     create table(:plans) do
-      add :area, :integer
+      add :area, :float
       add :bed_id, references(:beds, on_delete: :nothing)
       add :plant_id, references(:plants, on_delete: :nothing)
       add :user_id, references(:users, type: :id, on_delete: :delete_all)
@@ -12,7 +12,6 @@ defmodule Planter.Repo.Migrations.CreatePlans do
     end
 
     create index(:plans, [:user_id])
-
     create index(:plans, [:bed_id])
     create index(:plans, [:plant_id])
   end
